Renamed inputContainer to configuration
